feat(api): migrate attach child organization route
from lib to src/organizational-entities
