+++ api/src/organizational-entities/domain/usecases/index.js
@@ -1,0 +1,20 @@
+import { dirname, join } from 'node:path';
+import { fileURLToPath } from 'node:url';
+
+import { injectDependencies } from '../../../shared/infrastructure/utils/dependency-injection.js';
+import { importNamedExportsFromDirectory } from '../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
+import * as organizationForAdminRepository from '../../infrastructure/repositories/organization-for-admin-repository.js';
+
+const path = dirname(fileURLToPath(import.meta.url));
+
+const dependencies = {
+  organizationForAdminRepository,
+};
+
+const usecasesWithoutInjectedDependencies = {
+  ...(await importNamedExportsFromDirectory({ path: join(path, './'), ignoredFileNames: ['index.js'] })),
+};
+
+const usecases = injectDependencies(usecasesWithoutInjectedDependencies, dependencies);
+
+export { usecases };
