+++ api/src/organizational-entities/infrastructure/repositories/organization-for-admin-repository.js
@@ -1,11 +1,11 @@
 import _ from 'lodash';
 
-import { knex } from '../../../db/knex-database-connection.js';
-import { MissingAttributesError, NotFoundError } from '../../domain/errors.js';
-import { OrganizationInvitation } from '../../domain/models/OrganizationInvitation.js';
-import { OrganizationForAdmin } from '../../domain/models/organizations-administration/OrganizationForAdmin.js';
-import { Tag } from '../../domain/models/Tag.js';
-import { DomainTransaction } from '../DomainTransaction.js';
+import { knex } from '../../../../db/knex-database-connection.js';
+import { MissingAttributesError, NotFoundError } from '../../../../lib/domain/errors.js';
+import { OrganizationInvitation } from '../../../../lib/domain/models/OrganizationInvitation.js';
+import { Tag } from '../../../../lib/domain/models/Tag.js';
+import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
+import { OrganizationForAdmin } from '../../domain/models/OrganizationForAdmin.js';
 
 const ORGANIZATIONS_TABLE_NAME = 'organizations';
