+++ api/tests/integration/domain/usecases/update-certification-center-data-protection-officer-information_test.js
@@ -1,6 +1,6 @@
-import { DataProtectionOfficer } from '../../../../lib/domain/models/DataProtectionOfficer.js';
 import { updateCertificationCenterDataProtectionOfficerInformation } from '../../../../lib/domain/usecases/update-certification-center-data-protection-officer-information.js';
 import * as dataProtectionOfficerRepository from '../../../../lib/infrastructure/repositories/data-protection-officer-repository.js';
+import { DataProtectionOfficer } from '../../../../src/organizational-entities/domain/models/DataProtectionOfficer.js';
 import { databaseBuilder, expect } from '../../../test-helper.js';
 
 describe('Integration | UseCases | update-certification-center-data-protection-officer-information', function () {
