+++ api/tests/tooling/domain-builder/factory/build-data-protection-officer.js
@@ -1,4 +1,4 @@
-import { DataProtectionOfficer } from '../../../../lib/domain/models/DataProtectionOfficer.js';
+import { DataProtectionOfficer } from '../../../../src/organizational-entities/domain/models/DataProtectionOfficer.js';
 
 function _buildDataProtectionOfficer({
   id = 10001,
