+++ api/tests/unit/domain/usecases/create-certification-center_test.js
@@ -1,5 +1,5 @@
-import { DataProtectionOfficer } from '../../../../lib/domain/models/DataProtectionOfficer.js';
 import { createCertificationCenter } from '../../../../lib/domain/usecases/create-certification-center.js';
+import { DataProtectionOfficer } from '../../../../src/organizational-entities/domain/models/DataProtectionOfficer.js';
 import { domainBuilder, expect, sinon } from '../../../test-helper.js';
 
 describe('Unit | UseCase | create-certification-center', function () {
