+++ api/tests/unit/domain/usecases/organizations-administration/update-organization-information_test.js
@@ -1,6 +1,6 @@
 import { NotFoundError } from '../../../../../lib/domain/errors.js';
-import { OrganizationForAdmin } from '../../../../../lib/domain/models/organizations-administration/OrganizationForAdmin.js';
 import { usecases } from '../../../../../lib/domain/usecases/index.js';
+import { OrganizationForAdmin } from '../../../../../src/organizational-entities/domain/models/OrganizationForAdmin.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../../test-helper.js';
 const { updateOrganizationInformation } = usecases;
